flaskBiblioteca/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.1"
description = ""
authors = ["Milton-Luis <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
dynaconf = "^3.2.4"
flask-mail = "^0.9.1"
flask-admin = "^1.6.1"
flask-argon2 = "^0.3.0.0"
python-dotenv = "^1.0.0"
flask-jwt-extended = "^4.6.0"
flask-login = "^0.6.3"
flask-migrate = "^4.0.5"
bootstrap-flask = "^2.3.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flask-shell-ipython = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
